Factor out a generic PersonMixin for views
This factors out the code that gets a Person from a person_id in
kwargs, which is useful to many views, not just those for handling
editing of other names.
